refactor(baseball): 패키지명 변경 및 클래스 이동, 클래스 명 변경
