+++ src/test/java/baseball/controller/CountControllerTest.java
@@ -1,9 +1,9 @@
 package baseball.controller;
 
-import baseball.util.Count;
+import baseball.domain.Count;
+import baseball.domain.CountCalculator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -16,18 +16,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class CountControllerTest {
 
-    private CountController countController;
+    private CountCalculator countCalculator;
 
     @BeforeEach
     void setUp() {
-        countController = new CountController();
+        countCalculator = new CountCalculator();
     }
 
     @DisplayName("카운트 기능을 제대로 수행하는지 확인")
     @ParameterizedTest
     @MethodSource("provideNumbers")
     void generateCorrectCount(List<Integer> computerNumbers, List<Integer> userNumbers, int strike, int ball) {
-        Count result = countController.generateCount(computerNumbers, userNumbers);
+        Count result = countCalculator.generateCount(computerNumbers, userNumbers);
         Count answer = new Count(strike, ball);
 
         assertThat(result).usingRecursiveComparison().isEqualTo(answer);
@@ -44,7 +44,7 @@ class CountControllerTest {
     @ParameterizedTest
     @MethodSource("provideCompareNumbers")
     void checkTrueOrFalse(List<Integer> computerNumbers, String input, boolean answer) {
-        boolean result = countController.calculateCount(computerNumbers, input);
+        boolean result = countCalculator.calculateCount(computerNumbers, input);
 
         assertThat(result).isEqualTo(answer);
     }
